pyproject.toml
[tool.poetry]
name = "techecho"
version = "0.1.0"
description = ""
authors = [
    "alex <alexyongxinlin@gmail.com>",
    "buding <buding033171@gmail.com>",
    "chen weichen <newsforcwc@gmail.com>",
    "chi <a27976566@gmail.com>",
    "Eudora <eudoralin.tw@gmail.com>",
    "hong <sabina880731@gmail.com>",
    "tony <tony.t0705@gmail.com>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
line-pay = "^0.2.0"
psycopg2-binary = { version = "^2.9.9", markers = "sys_platform == 'linux' or sys_platform == 'darwin'" }
psycopg2 = { version = "^2.9.9", markers = "sys_platform == 'win32'" }
django-taggit = "^6.0.0"
django-htmx = "^1.19.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
daphne = "^4.1.2"
django-allauth = "^64.2.1"
pyjwt = "^2.9.0"
pillow = "^10.4.0"
django-storages = "^1.14.4"
boto3 = "^1.35.19"
markdown = "^3.7"
markdown2 = "^2.5.0"
channels-redis = "^4.2.0"



[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
pre-commit = "^3.8.0"
python-dotenv = "^1.0.1"
commitizen = "^3.29.0"
faker = "^28.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"